Kali uses msf to invade Windows computers
Kali uses msf to invade Windows computers (the most detailed penetration process, generating Trojan horses, and monitoring screens).
Legal Disclaimer: This content is intended solely for reference or educational purposes. It is explicitly stated that any utilization of this information for illegal or unauthorized activities is strictly prohibited and may result in legal consequences.